+++ data/src/main/java/com/wakeup/data/repository/MomentRepositoryImpl.kt
@@ -37,9 +37,18 @@ class MomentRepositoryImpl @Inject constructor(
             }
         }
 
+    override fun getAllMoments(): Flow<List<Moment>> =
+        localDataSource.getAllMoments().map { momentInfoList ->
+            momentInfoList.map { momentInfo ->
+                momentInfo.toDomain(util.getPictureInInternalStorage(momentInfo.pictures,
+                    momentInfo.moment.thumbnailId))
+            }
+        }
+
     override suspend fun saveMoment(moment: Moment) {
         val globeIndex = localDataSource.getGlobeId(moment.globes.first().name)
 
+        // 사진 없다면, 즉시 모먼트 저장
         if (moment.pictures.isEmpty()) {
             val momentIndex =
                 localDataSource.saveMoment(moment.toEntity(moment.place, null))
@@ -49,15 +58,20 @@ class MomentRepositoryImpl @Inject constructor(
             return
         }
 
+        // 1. 내부 저장소에 이미지 저장
         val pictureFileNames = util.savePictureInInternalStorageAndGetFileName(moment.pictures)
         val pictureIndexes = localDataSource.savePictures(pictureFileNames)
 
+        // 2. 모먼트 DB 저장
         val momentIndex =
             localDataSource.saveMoment(moment.toEntity(moment.place, pictureIndexes.first()))
 
+        // 3. [모먼트 - 사진] DB 저장
         localDataSource.saveMomentPictures(pictureIndexes.map { pictureId ->
             MomentPictureXRef(momentId = momentIndex, pictureId = pictureId)
         })
+
+        // 4. [모먼트 - 글로브] DB 저장
         localDataSource.saveMomentGlobe(
             MomentGlobeXRef(momentId = momentIndex, globeId = globeIndex)
         )
