+++ domain/src/main/java/com/wakeup/domain/repository/MomentRepository.kt
@@ -1,10 +1,8 @@
 package com.wakeup.domain.repository
 
 import androidx.paging.PagingData
-import com.wakeup.domain.model.Picture
 import com.wakeup.domain.model.Location
 import com.wakeup.domain.model.Moment
-import com.wakeup.domain.model.Place
 import com.wakeup.domain.model.SortType
 import kotlinx.coroutines.flow.Flow
 
@@ -14,6 +12,8 @@ interface MomentRepository {
         query: String,
         myLocation: Location? = null
     ): Flow<PagingData<Moment>>
+
+    fun getAllMoments(): Flow<List<Moment>>
 
     suspend fun saveMoment(moment: Moment)
 }
