+++ presentation/src/main/java/com/wakeup/presentation/ui/home/map/MapFragment.kt
@@ -4,18 +4,24 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.google.android.material.snackbar.Snackbar
+import androidx.navigation.fragment.findNavController
 import com.naver.maps.map.LocationTrackingMode
 import com.naver.maps.map.NaverMap
 import com.naver.maps.map.OnMapReadyCallback
+import com.naver.maps.map.overlay.Marker
+import com.naver.maps.map.overlay.Overlay.OnClickListener
 import com.naver.maps.map.util.FusedLocationSource
 import com.wakeup.presentation.databinding.FragmentMapBinding
+import com.wakeup.presentation.extension.getFadeInAnimator
 import com.wakeup.presentation.model.LocationModel
+import com.wakeup.presentation.model.MomentModel
+import com.wakeup.presentation.ui.home.HomeFragmentDirections
 import com.wakeup.presentation.ui.home.HomeViewModel
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
@@ -32,7 +38,7 @@ class MapFragment : Fragment(), OnMapReadyCallback {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
-        savedInstanceState: Bundle?
+        savedInstanceState: Bundle?,
     ): View {
         binding = FragmentMapBinding.inflate(inflater, container, false)
         binding.lifecycleOwner = viewLifecycleOwner
@@ -41,9 +47,20 @@ class MapFragment : Fragment(), OnMapReadyCallback {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
+        initMomentPreviewClickListener()
         initMap()
         collectData()
+    }
+
+    // 프리뷰 터치시, 상세 페이지 이동
+    private fun initMomentPreviewClickListener() {
+        binding.momentPreview.root.setOnClickListener {
+            binding.momentPreview.momentModel?.let {
+                val action =
+                    HomeFragmentDirections.actionMapFragmentToMomentDetailFragment(it)
+                findNavController().navigate(action)
+            }
+        }
     }
 
     private fun initMap() {
@@ -74,6 +91,11 @@ class MapFragment : Fragment(), OnMapReadyCallback {
     override fun onMapReady(naverMap: NaverMap) {
         this.naverMap = naverMap
 
+        // 지도 터치시, 정보 창 사라짐 설정
+        mapHelper.setViewFadeOutClickListener(naverMap,
+            binding.momentPreview.root,
+            MOMENT_PREVIEW_ANIM_DURATION)
+
         // 위치 관련 설정
         mapHelper.apply {
             setCurrentLocation(naverMap, locationSource)
@@ -82,12 +104,52 @@ class MapFragment : Fragment(), OnMapReadyCallback {
             setLogoView(naverMap, binding.lvLogo)
         }
 
+        // 전체 모먼트 불러오기
+        // TODO 속도가 엄청 느려진다.
+        viewLifecycleOwner.lifecycleScope.launch {
+            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
+                viewModel.allMoments.collectLatest { moments ->
+                    moments.forEach { momentModel ->
+                        setMarkerClickListener(momentModel)
+                    }
+                }
+            }
+        }
+
+        // Paging 있이, Moment 가져오기
+        /*viewLifecycleOwner.lifecycleScope.launch {
+            momentAdapter.loadStateFlow.collectLatest {
+                momentAdapter.snapshot().items.forEach { moment ->
+                    setMarker(moment)
+                }
+            }
+        }*/
+
         // 테스트 마커 설정
-        mapHelper.setTestMarker(naverMap) {
-            Snackbar.make(binding.root, "${it.tag}번째 마커", Snackbar.LENGTH_SHORT).show()
-            mapHelper.setDarkMode(naverMap)
+        /* mapHelper.setTestMarker(naverMap) {
+             Snackbar.make(binding.root, "${it.tag}번째 마커", Snackbar.LENGTH_SHORT).show()
+             mapHelper.setDarkMode(naverMap)
+             true
+         } */
+    }
+
+    private fun setMarkerClickListener(moment: MomentModel) {
+        val clickListener = OnClickListener { marker ->
+
+            (marker as Marker).apply {
+                mapHelper.setMarkerFocused(this)
+                mapHelper.moveCamera(naverMap, position)
+                binding.momentModel = (tag as MomentModel)
+            }
+
+            binding.momentPreview.root.let {
+                it.isVisible = true
+                it.getFadeInAnimator(MapFragment.MOMENT_PREVIEW_ANIM_DURATION).start()
+            }
             true
         }
+
+        mapHelper.setMarker(naverMap, moment, clickListener)
     }
 
     // Deprecated 되었지만,
@@ -116,5 +178,6 @@ class MapFragment : Fragment(), OnMapReadyCallback {
 
     companion object {
         const val LOCATION_PERMISSION_REQUEST_CODE = 1000
+        const val MOMENT_PREVIEW_ANIM_DURATION = 300L
     }
 }
