--- presentation/src/main/java/com/wakeup/presentation/ui/map/MapFragment.kt
@@ -1,326 +1,0 @@
-package com.wakeup.presentation.ui.map
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.view.animation.AnimationUtils
-import android.view.inputmethod.EditorInfo
-import android.widget.ArrayAdapter
-import androidx.core.view.isVisible
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.RecyclerView
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.textfield.MaterialAutoCompleteTextView
-import com.naver.maps.map.LocationTrackingMode
-import com.naver.maps.map.NaverMap
-import com.naver.maps.map.OnMapReadyCallback
-import com.naver.maps.map.overlay.Marker
-import com.naver.maps.map.overlay.Overlay.OnClickListener
-import com.naver.maps.map.util.FusedLocationSource
-import com.wakeup.domain.model.SortType
-import com.wakeup.presentation.R
-import com.wakeup.presentation.adapter.MomentPagingAdapter
-import com.wakeup.presentation.databinding.BottomSheetBinding
-import com.wakeup.presentation.databinding.FragmentMapBinding
-import com.wakeup.presentation.extension.getFadeInAnimator
-import com.wakeup.presentation.extension.getNavigationResultFromTop
-import com.wakeup.presentation.model.LocationModel
-import com.wakeup.presentation.model.MomentModel
-import com.wakeup.presentation.ui.MainActivity
-import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.launch
-
-@AndroidEntryPoint
-class MapFragment : Fragment(), OnMapReadyCallback {
-    private lateinit var binding: FragmentMapBinding
-    private val viewModel: MapViewModel by viewModels()
-    private val momentAdapter = MomentPagingAdapter()
-
-    private lateinit var naverMap: NaverMap
-    private lateinit var locationSource: FusedLocationSource
-    private lateinit var mapHelper: MapHelper
-
-    private var scrollToTop = false
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        binding = FragmentMapBinding.inflate(inflater, container, false)
-        binding.lifecycleOwner = viewLifecycleOwner
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        initMomentPreviewClickListener()
-
-        initMap()
-
-        initBottomSheet()
-        setAdapterListener()
-        setSearchBarListener()
-
-        collectMoments()
-        updateMoments()
-    }
-
-    // 프리뷰 터치시, 상세 페이지 이동
-    private fun initMomentPreviewClickListener() {
-        binding.momentPreview.root.setOnClickListener {
-            binding.momentPreview.momentModel?.let {
-                val action =
-                    MapFragmentDirections.actionMapFragmentToMomentDetailFragment(it)
-                findNavController().navigate(action)
-            }
-        }
-    }
-
-    private fun initMap() {
-        mapHelper = MapHelper(requireContext())
-
-        mapHelper.initMap(childFragmentManager, this)
-
-        locationSource =
-            FusedLocationSource(this, LOCATION_PERMISSION_REQUEST_CODE).apply {
-                // 사용자가 바라보는 방향으로 지도 회전 활성화
-                this.isCompassEnabled = true
-            }
-    }
-
-    private fun initBottomSheet() {
-        with(binding.bottomSheet) {
-            setState(this)
-            setMenus(this)
-            setAdapter(this)
-            setCallback(this)
-        }
-    }
-
-    private fun setState(binding: BottomSheetBinding) {
-        val behavior = BottomSheetBehavior.from(binding.bottomSheet)
-        behavior.state = viewModel.bottomSheetState.value
-    }
-
-    private fun setMenus(binding: BottomSheetBinding) {
-        val items = listOf(
-            getString(R.string.most_recent),
-            getString(R.string.oldest),
-            getString(R.string.nearest)
-        )
-        val menuAdapter = ArrayAdapter(requireContext(), R.layout.item_menu, items)
-        binding.textField.viewTreeObserver.addOnGlobalLayoutListener {
-            (binding.textField.editText as? MaterialAutoCompleteTextView)?.setAdapter(menuAdapter)
-        }
-
-        binding.sortMenu.setOnItemClickListener { _, _, position, _ ->
-            viewModel.location.value = null
-            viewModel.sortType.value = when (position) {
-                0 -> SortType.MOST_RECENT
-                1 -> SortType.OLDEST
-                else -> {
-                    locationSource.lastLocation?.apply {
-                        viewModel.location.value = LocationModel(latitude, longitude)
-                    }
-                    SortType.NEAREST
-                }
-            }
-            viewModel.fetchMoments()
-        }
-    }
-
-    private fun setAdapter(binding: BottomSheetBinding) {
-        binding.rvMoments.adapter = momentAdapter
-    }
-
-    private fun setCallback(binding: BottomSheetBinding) {
-        val behavior = BottomSheetBehavior.from(binding.bottomSheet)
-        if (behavior.state == BottomSheetBehavior.STATE_COLLAPSED) {
-            binding.textField.visibility = View.INVISIBLE
-        }
-
-        behavior.addBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
-            override fun onStateChanged(bottomSheet: View, newState: Int) {
-                when (newState) {
-                    BottomSheetBehavior.STATE_EXPANDED -> {
-                        changeVisibleMenu(true)
-                        viewModel.bottomSheetState.value = BottomSheetBehavior.STATE_EXPANDED
-                    }
-                    BottomSheetBehavior.STATE_COLLAPSED -> {
-                        changeVisibleMenu(false)
-                        viewModel.bottomSheetState.value = BottomSheetBehavior.STATE_COLLAPSED
-                    }
-                }
-            }
-
-            override fun onSlide(bottomSheet: View, slideOffset: Float) = Unit
-        })
-    }
-
-    private fun changeVisibleMenu(isVisible: Boolean) = with(binding.bottomSheet.textField) {
-        if (isVisible) {
-            visibility = View.VISIBLE
-            animation = AnimationUtils.loadAnimation(context, R.anim.fade_in)
-        } else {
-            visibility = View.INVISIBLE
-            animation = AnimationUtils.loadAnimation(context, R.anim.fade_out)
-        }
-    }
-
-    private val adapterDataObserver = object : RecyclerView.AdapterDataObserver() {
-
-        override fun onItemRangeInserted(positionStart: Int, itemCount: Int) {
-            super.onItemRangeInserted(positionStart, itemCount)
-            if (scrollToTop) {
-                binding.bottomSheet.rvMoments.scrollToPosition(0)
-            }
-            scrollToTop = false
-        }
-
-        override fun onItemRangeMoved(fromPosition: Int, toPosition: Int, itemCount: Int) {
-            super.onItemRangeMoved(fromPosition, toPosition, itemCount)
-            binding.bottomSheet.rvMoments.scrollToPosition(0)
-        }
-    }
-
-    private fun setAdapterListener() {
-        momentAdapter.registerAdapterDataObserver(adapterDataObserver)
-
-        momentAdapter.addLoadStateListener {
-            binding.bottomSheet.hasMoments = momentAdapter.itemCount > 0
-        }
-    }
-
-    private fun setSearchBarListener() {
-        binding.ivMenu.setOnClickListener {
-            (activity as MainActivity).openNavDrawer()
-        }
-
-        binding.etSearch.setOnEditorActionListener { textView, i, _ ->
-            if (i == EditorInfo.IME_ACTION_SEARCH) {
-                scrollToTop = true
-                viewModel.setSearchQuery(textView.text.toString())
-                viewModel.fetchMoments()
-
-                (activity as MainActivity).hideKeyboard()
-            }
-            false // true: 계속 search 가능
-        }
-    }
-
-    private fun collectMoments() {
-        viewLifecycleOwner.lifecycleScope.launch {
-            viewLifecycleOwner.repeatOnLifecycle(Lifecycle.State.STARTED) {
-                viewModel.moments.collectLatest {
-                    momentAdapter.submitData(it)
-                }
-            }
-        }
-    }
-
-    private fun updateMoments() {
-        findNavController().getNavigationResultFromTop<Boolean>("isUpdated")
-            ?.observe(viewLifecycleOwner) { isUpdated ->
-                scrollToTop = isUpdated
-                if (isUpdated) {
-                    binding.bottomSheet.sortMenu.setText(R.string.most_recent)
-                    viewModel.sortType.value = SortType.MOST_RECENT
-                    viewModel.fetchMoments()
-                }
-            }
-    }
-
-    override fun onMapReady(naverMap: NaverMap) {
-        this.naverMap = naverMap
-
-        // 지도 터치시, 정보 창 사라짐 설정
-        mapHelper.setViewFadeOutClickListener(naverMap,
-            binding.momentPreview.root,
-            MOMENT_PREVIEW_ANIM_DURATION)
-
-        // 위치 관련 설정
-        mapHelper.apply {
-            setCurrentLocation(naverMap, locationSource)
-            setLocationButtonView(naverMap, binding.lbvLocation)
-            setScaleBarView(naverMap, binding.sbvScale)
-            setLogoView(naverMap, binding.lvLogo)
-        }
-
-        // TODO Paging 없이, Moment 가져오기
-        viewLifecycleOwner.lifecycleScope.launch {
-            momentAdapter.loadStateFlow.collectLatest {
-                momentAdapter.snapshot().items.forEach { moment ->
-                    setMarkerClickListener(moment)
-                }
-            }
-        }
-
-        // 테스트 마커 설정
-        /* mapHelper.setTestMarker(naverMap) {
-             Snackbar.make(binding.root, "${it.tag}번째 마커", Snackbar.LENGTH_SHORT).show()
-             mapHelper.setDarkMode(naverMap)
-             true
-         } */
-    }
-
-    private fun setMarkerClickListener(moment: MomentModel) {
-        val clickListener = OnClickListener { marker ->
-
-            (marker as Marker).apply {
-                mapHelper.setMarkerFocused(this)
-                mapHelper.moveCamera(naverMap, position)
-                binding.momentModel = (tag as MomentModel)
-            }
-
-            binding.momentPreview.root.let {
-                it.isVisible = true
-                it.getFadeInAnimator(MOMENT_PREVIEW_ANIM_DURATION).start()
-            }
-            true
-        }
-
-        mapHelper.setMarker(naverMap, moment, clickListener)
-    }
-
-    // Deprecated 되었지만,
-    // Location, Camera Tracking을 위해 Naver의 FusedLocationSource을 사용하려면 어쩔 수 없음
-    @Deprecated("Deprecated in Java")
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray,
-    ) {
-        if (locationSource.onRequestPermissionsResult(
-                requestCode, permissions,
-                grantResults
-            )
-        ) {
-            if (!locationSource.isActivated) { // 권한 거부됨
-                naverMap.locationTrackingMode = LocationTrackingMode.None
-            }
-
-            naverMap.locationTrackingMode = LocationTrackingMode.Follow
-
-            return
-        }
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-    }
-
-    override fun onDestroyView() {
-        momentAdapter.unregisterAdapterDataObserver(adapterDataObserver)
-        binding.unbind()
-        super.onDestroyView()
-    }
-
-    companion object {
-        const val LOCATION_PERMISSION_REQUEST_CODE = 1000
-        const val MOMENT_PREVIEW_ANIM_DURATION = 300L
-    }
-}
